Overhaul path selection; Enable nullable for InstallerSettings
Path selection now occurs within the main window view model instead of
creating a useless view which just serves to provide async support. As
a result, the class has turned from a ViewModel into a static utility
class. In addition, nullable reference types are now enabled for
InstallerSettings, with the idea being to spread this to the rest of the
project in the future.